pyproject.toml
[project]
name = "travelsync-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "autoflake>=2.3.3",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "ruff>=0.15.10",
    "sqlalchemy>=2.0.48",
    "starlette[full]>=0.52.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
